changed to double new attempt

+++ src/main/java/com/example/springdocker/MyMathCalc.java
@@ -5,11 +5,12 @@ public class MyMathCalc {
     public int add(int a, int b){
         return a + b;
     }
+  
     public int sub(int a, int b){
         return a - b;
     }
-
-    public float div(int a, int b){
+  
+    public double div(double a, double b){
         return a / b;
 
     }
